requirements.txt
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python-ng==2.2.3
MarkupSafe==3.0.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pymongo==4.10.1
Werkzeug==3.0.4
